ref: moved Invoice schema, types and utils to /schemas/invoice
