+++ functions/invoices/getInvoiceBase64.ts
@@ -1,5 +1,5 @@
 import GET from '@/api/GET'
-import { Invoice } from '@/schemas/Invoice'
+import { Invoice } from '@/schemas/invoice/Invoice'
 
 export default async function getInvoiceBase64(invoice_id: Invoice['system_id']) {
   const { pdf_base64_encoded } = await GET<{ pdf_base64_encoded: string }>(`invoices/${invoice_id}/pdf`, ['locale=de_DE'])
