+++ functions/invoices/getInvoices.ts
@@ -1,7 +1,7 @@
 import GET from '@/api/GET'
 import parseRawInvoice from '@/functions/invoices/parseRawInvoice'
-import { Invoice } from '@/schemas/Invoice'
 import { RawInvoices } from '@/schemas/invoice/RawInvoices'
+import { Invoice } from '@/schemas/invoice/Invoice'
 
 export async function getInvoices(limit: number = 1000) {
   if (limit === 0) return []
