+++ jest.mocks.ts
@@ -1,5 +1,5 @@
 import { getAuthorization } from '@/config/authorization'
-import { getDummyInvoice } from '@/schemas/Invoice'
+import { getDummyInvoice } from '@/schemas/invoice/Invoice'
 
 jest.mock('@/functions/invoices/getInvoices', () => ({
   getInvoices: jest.fn((limit: number) => {
